chore($formatting): enforce consistent formatting
Leverage an `.editorconfig` file to enforce consistent formatting in repo. Spaces (versus tabs) are
chosen because tabs can vary in meaning across different OS's; compared to spaces which have the
same meaning. This setting can be updated to tabs if desired. Important setting is that each file
contains a new_line at the end. This is required for git gui and other git tools for committing
specific lines versus hunks/entire file.