fix($tests): enable runnings tests via wallabyjs
The `.babelrc` file is the configuration that babel will use. WallabyJS is configured to use babel to process/compile JavaScript files, but without the babel configuration, babel does not know what or how to process the files. Since we are not using React and only using ES2015 language features, we use the `babel-preset-es2015` preset in the `.babelrc`. This is an NPM package that is now a dev dependency. It represents a collection of preset configurations for babel, culminating to the ES2015 specification feature set.

The `babel` NPM package no longer exists; instead Babel consists of a collection of packages that can be installed depending on what you need to use versus an all or nothing approach. For example, if you want to run Babel from the CLI, you would need to install `babel-cli`. The core of babel which wallaby will require is `babel-core`.

The `setup()` function of wallaby’s configuration file enables setup code to be run before any tests. Functions such as `describe`, `it`, `beforeEach`, etc. are provided by the testing framework, in this case Mocha. However, `expect` is provided separately from the testing framework. This is done so consumers of Mocha can choose what assertion library they want to use. We are using `chai` as our assertion library. However, we called `expect()` in our tests as if it is globally available. This is not globally available by default. Therefore, we can either `import {expect} from ‘chai’;` in each of our tests OR we can set `global.expect` equal to this value. This essentially adds an `expect` function into the global space that can be used in all tests. It is done in the setup function of wallaby’s configuration to ensure it is available on all tests.